migrated dtw functionality to sequence module. fixes #715
