+++ librosa/core/__init__.py
@@ -92,8 +92,8 @@ Pitch and tuning
     pitch_tuning
     piptrack
 
-Dynamic Time Warping
---------------------
+Deprecated (moved)
+------------------
 .. autosummary::
     :toctree: generated/
 
@@ -106,7 +106,13 @@ from .audio import *  # pylint: disable=wildcard-import
 from .spectrum import *  # pylint: disable=wildcard-import
 from .pitch import *  # pylint: disable=wildcard-import
 from .constantq import *  # pylint: disable=wildcard-import
-from .dtw import *  # pylint: disable=wildcard-import
 from .harmonic import *  # pylint: disable=wildcard-import
+
+from ..util.decorators import moved as _moved
+from ..util import fill_off_diagonal as _fod
+from ..sequence import dtw as _dtw
+
+dtw = _moved('librosa.sequence.dtw', '0.6.1', '0.7')(_dtw)
+fill_off_diagonal = _moved('librosa.util.fill_off_diagonal', '0.6.1', '0.7')(_fod)
 
 __all__ = [_ for _ in dir() if not _.startswith('_')]
