+++ tests/test_dtw.py
@@ -17,7 +17,7 @@ warnings.simplefilter('always')
 def test_1d_input():
     X = np.array([[1], [3], [3], [8], [1]])
     Y = np.array([[2], [0], [0], [8], [7], [2]])
-    librosa.dtw(X=X, Y=Y)
+    librosa.sequence.dtw(X=X, Y=Y)
 
 
 def test_dtw_global():
@@ -32,11 +32,11 @@ def test_dtw_global():
                      [9., 11., 13., 7., 8., 14.],
                      [10, 10., 11., 14., 13., 9.]])
 
-    mut_D, _ = librosa.dtw(X, Y)
+    mut_D, _ = librosa.sequence.dtw(X, Y)
     assert np.array_equal(gt_D, mut_D)
 
     # Check that it works without backtracking
-    mut_D2 = librosa.dtw(X, Y, backtrack=False)
+    mut_D2 = librosa.sequence.dtw(X, Y, backtrack=False)
     assert np.array_equal(mut_D, mut_D2)
 
 
@@ -53,7 +53,7 @@ def test_dtw_global_constrained():
                      [np.inf, np.inf, 13., 7., 8., 14.],
                      [np.inf, np.inf, np.inf, 14., 13., 9.]])
 
-    mut_D = librosa.dtw(X, Y, backtrack=False, global_constraints=True, band_rad=0.5)
+    mut_D = librosa.sequence.dtw(X, Y, backtrack=False, global_constraints=True, band_rad=0.5)
     assert np.array_equal(gt_D, mut_D)
 
 
@@ -74,19 +74,19 @@ def test_dtw_global_supplied_distance_matrix():
 
     # Supply precomputed distance matrix and specify an invalid distance
     # metric to verify that it isn't used.
-    mut_D, _ = librosa.dtw(C=C, metric='invalid')
+    mut_D, _ = librosa.sequence.dtw(C=C, metric='invalid')
 
     assert np.array_equal(gt_D, mut_D)
 
 
 @raises(librosa.ParameterError)
 def test_dtw_incompatible_args_01():
-    librosa.dtw(C=1, X=1, Y=1)
+    librosa.sequence.dtw(C=1, X=1, Y=1)
 
 
 @raises(librosa.ParameterError)
 def test_dtw_incompatible_args_02():
-    librosa.dtw(C=None, X=None, Y=None)
+    librosa.sequence.dtw(C=None, X=None, Y=None)
 
 
 
@@ -94,21 +94,21 @@ def test_dtw_incompatible_args_02():
 def test_dtw_incompatible_sigma_add():
     X = np.array([[1, 3, 3, 8, 1]])
     Y = np.array([[2, 0, 0, 8, 7, 2]])
-    librosa.dtw(X=X, Y=Y, weights_add=np.arange(10))
+    librosa.sequence.dtw(X=X, Y=Y, weights_add=np.arange(10))
 
 
 @raises(librosa.ParameterError)
 def test_dtw_incompatible_sigma_mul():
     X = np.array([[1, 3, 3, 8, 1]])
     Y = np.array([[2, 0, 0, 8, 7, 2]])
-    librosa.dtw(X=X, Y=Y, weights_mul=np.arange(10))
+    librosa.sequence.dtw(X=X, Y=Y, weights_mul=np.arange(10))
 
 
 @raises(librosa.ParameterError)
 def test_dtw_incompatible_sigma_diag():
     X = np.array([[1, 3, 3, 8, 1, 2]])
     Y = np.array([[2, 0, 0, 8, 7]])
-    librosa.dtw(X=X, Y=Y, step_sizes_sigma=np.ones((1, 2), dtype=int))
+    librosa.sequence.dtw(X=X, Y=Y, step_sizes_sigma=np.ones((1, 2), dtype=int))
 
 
 def test_dtw_global_diagonal():
@@ -118,7 +118,7 @@ def test_dtw_global_diagonal():
 
     gt_wp = list(zip(list(range(10)), list(range(10))))[::-1]
 
-    mut_D, mut_wp = librosa.dtw(X, Y, subseq=True, metric='cosine',
+    mut_D, mut_wp = librosa.sequence.dtw(X, Y, subseq=True, metric='cosine',
                                 step_sizes_sigma=np.array([[1, 1]]),
                                 weights_mul=np.array([1, ]))
 
@@ -136,7 +136,7 @@ def test_dtw_subseq():
     noise = np.random.rand(noise_len)
     Y = np.concatenate((noise, noise, X, noise))
 
-    _, mut_wp = librosa.dtw(X, Y, subseq=True)
+    _, mut_wp = librosa.sequence.dtw(X, Y, subseq=True)
 
     # estimated sequence has to match original sequence
     # note the +1 due to python indexing
@@ -151,49 +151,8 @@ def test_dtw_subseq_sym():
     gt_wp_XY = np.array([[3, 5], [2, 4], [1, 3], [0, 2]])
     gt_wp_YX = np.array([[5, 3], [4, 2], [3, 1], [2, 0]])
 
-    _, mut_wp_XY = librosa.dtw(X, Y, subseq=True)
-    _, mut_wp_YX = librosa.dtw(Y, X, subseq=True)
+    _, mut_wp_XY = librosa.sequence.dtw(X, Y, subseq=True)
+    _, mut_wp_YX = librosa.sequence.dtw(Y, X, subseq=True)
 
     assert np.array_equal(gt_wp_XY, mut_wp_XY)
     assert np.array_equal(gt_wp_YX, mut_wp_YX)
-
-
-def test_dtw_fill_off_diagonal_8_8():
-    # Case 1: Square matrix (N=M)
-    mut_x = np.ones((8, 8))
-    librosa.fill_off_diagonal(mut_x, 0.25)
-
-    gt_x = np.array([[1, 1, 0, 0, 0, 0, 0, 0],
-                     [1, 1, 1, 0, 0, 0, 0, 0],
-                     [0, 1, 1, 1, 0, 0, 0, 0],
-                     [0, 0, 1, 1, 1, 0, 0, 0],
-                     [0, 0, 0, 1, 1, 1, 0, 0],
-                     [0, 0, 0, 0, 1, 1, 1, 0],
-                     [0, 0, 0, 0, 0, 1, 1, 1],
-                     [0, 0, 0, 0, 0, 0, 1, 1]])
-
-    assert np.array_equal(mut_x, gt_x)
-    assert np.array_equal(mut_x, gt_x.T)
-
-
-def test_dtw_fill_off_diagonal_8_12():
-    # Case 2a: N!=M
-    mut_x = np.ones((8, 12))
-    librosa.fill_off_diagonal(mut_x, 0.25)
-
-    gt_x = np.array([[1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0],
-                     [1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0],
-                     [0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0],
-                     [0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0],
-                     [0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0],
-                     [0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1, 0],
-                     [0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1],
-                     [0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1]])
-
-    assert np.array_equal(mut_x, gt_x)
-
-    # Case 2b: (N!=M).T
-    mut_x = np.ones((8, 12)).T
-    librosa.fill_off_diagonal(mut_x, 0.25)
-
-    assert np.array_equal(mut_x, gt_x.T)
